changing modal.confirm to a modal as we are able to set the visibility
